algo: changed how we accept graph merges
before I used a simple comparison of compression ratio before and after
this was much too strict as it would truncate the merge algorithm too
early (leaves of the tree are not complex by definition).
now, we compare the amount of sequence merged to the number of blocks
created in the process.
gamma sets the length scale to fix units.