Feat: native distance computation
To remove the hard dependency on mash, we include a distance computation
based on minimizer overlaps. This should, in principle, be more in line
with our underlying alignment algorithm. This required some refactoring
of how ordering() is called within align.jl. Additionally, a new command
line argument is recognized that allows one to toggle to the old
computation if desired.