Prototype of distance computation
There is a significant problem with the algorithm as written now.
Specifically, we visit each block once, and look at the left/right
breakpoints (taking into account strandedness). For all such
breakpoints, we analyze to see if the edge that takes individuals from
the neighboring block to the current block is equivalent. If not, we log
the breakpoint. This almost does the right thing, but for duplicated
blocks (where an individual shows up multiple times) it doesn't consider
that we only need to find the conjugate breakpoint *once* per pair.
Instead, this will erroneously log a breakpoint for most duplications.