Feat: added command line options to build
Added a flag to pangraph build to allow for tunable sensitivity
in the use of minimap. Additionally, exposed the maximum number
of iterations done during self-mapping to the user. The
documentation has been updated to reflect this.

Fixed a bug with our use of PyCall that lead to stale references during
compilation. The effect was a rare segfault during the last phase of
compilation.

Simpified usage of Node constructor.

Various small changes with alignment kernel.