feat: in-thread errors are handled by error channel
alignment tasks that run in parallel have a try-catch statement.
If an error occurs, it is passed to an error channel, together with
the stacktrace. The main thread waits on this error channel,
and logs any error received and throws a new error.
The error channel is closed in the last alignment taks,
when the root graph is created.