chore(ci): release container image when master branch is updated
So far, we pushed a docker image to DockerHub only when a git tag (a new version) was pushed.

Here I configure CI such that when pushed to master branch, the image with a special docker tag `:master` is also pushed to DockerHub.

That is, as before, when pushing a tag `1.2.3` the 2 docker tags are released:

```
neherlab/pangenome:1.2.3
neherlab/pangenome:latest
```

This way one can pull and run the latest version of the image  with

```
docker run ... neherlab/pangenome:latest
```

or, by omitting the default `:latest` tag:

```
docker run ... neherlab/pangenome
```

Or one could request a specific version with:

```
docker run ... neherlab/pangenome:1.2.3
```

Or one of the previous versions:

```
docker run ... neherlab/pangenome:1.1.1
```

But also, when pushing or merging a PR to master, an image is built from the current master branch and pushed to DockerHub:

```
docker run ... neherlab/pangenome:master
```

This images updates continuously as `master` branch updates and is useful for testing things manually before a tagged release.

The untagged builds from `master` don't update the `:latest` docker tag, so most end users are not affected by these builds. But the most curious can still try out the latest features from master if they want to, bu using the `:master` docker tag.