chore(ci): ensure branch builds are not cancelled by tag builds
Currently build triggered by a push to master is cancelled by the followup git tag on the same commit, due to `github.ref` being the same.

But we want `master` branch to deploy to `:master` docker tag, and git tags to be deployed to `:lastest`, so we need both.

To solve this, here I add  `github.ref_type` (`tag` or `branch`) to the `concurrency.group`, such that tag- and branch-triggered builds don't cancel each other.