feat: add `--verify` flag
This has a cost: Previously fasta content was consumed (by value) by the newly created blocks. Now, in order to verify we need to read fasta again, but it's gone. So when the `--verify` flag is true, we have to clone the fasta and retain it until the end of the run. This doubles sequence storage memory requirements.

Alternative approaches to consider:
- re-read fasta file(s) before verification - this way we don't need to store any additional data while `build`ing
- implement a separate command `pangraph verify` and re-read graph and fasta from disk/stdio - this way `build` and `verify` steps are separate. Perhaps it can even be called `reconstruct` and becomes a useful feature when users need to reconstruct sequences from the graph alone (e.g. if they are not the authors of the graph and only have the graph on their hands)