chore: generate C bindings for minimap crate only into output dir
In `minimap-sys` crate: this removes hardcoded `bindings.rs` in favor of always dynamically rebuilding them with `rust-bindgen`. The `vendored` feature of `minimap-sys` is removed and is always enabled now.

This is required to avoid clashes when compiling for different platforms and operating systems - the bindings are generated slightly differently on different setups.

It's a [best practice](https://rust-lang.github.io/rust-bindgen/tutorial-3.html) anyways.