feat: use ASCII char and special container for sequences
This introduces `AsciiChar` and `Seq` classes which represent an ASCII-character and ASCII-string. It's a `u8` and a `Vec` underneath, so it combines useful methods which are string-like and array-like for our application.

The main idea is that because Rust `String` are unicode and `char` are 32-bit integers, it's wasteful and slow to use that to represent genetic sequences. We should do better with `u8` 8-bit charters and with ASCII set assumption, e.g. no more unnecessary code splitting preservation, and we can afford more flexible and more efficient methods.

In reality, compared to the same transition in treetime project, where we observed 2x speedup, this did not introduce any significant speedup in my measurements in pangraph. However, memory occupied by sequences should reduce by a factor of 4.

Both of the new classes could use some more unit tests.