feat: make error message in mmseqs check prettier
This follows `color_eyre` conventions and uses different sections to make error message prettier.

It's important to note that any error can manifest with this exact message (e.g. if mmseqs requires libraries and they are missing, or if mmseqs binary is not executable, if it segfaults etc.), so I made the suggestion message a little more generic (not assuming that the error is "file not found").

![Image](https://github.com/user-attachments/assets/346bf3f9-db27-4c7f-a2d1-cd642c0d008f)