feat: make error message in mmseqs run prettier
This follows `color_eyre` conventions and uses different sections to make error message prettier.

I think that chaining `.arg()` and then adding `-k` with `.args()` does not necessarily actually adds `-k`, but I don't know how to check it. So I opted for manually constructing all-in-one array of args, which I pass using a single `.args()` call. This also allow us to print the full command to the user and inspect it during debugging.

Also refactored and simplified things a little.

![Image](https://github.com/user-attachments/assets/726a580f-c7f2-4e59-8f45-50089ab8125e)