feat: parse fasta description separately
This reads fasta sequence id (`.seq_name`) and description which comes after the first whitespace (`.desc`) separately.

The corresponding optional field (`.desc`) is added to the `FastaRecord` and `PangraphPath`.

The description is now emitted into reconstructed fasta where appropriate (perhaps there are more places where description can be written? I could have missed some) as well as into pangraph JSON (JSON Schema has been automatically adjusted).

A few things to consider:

 1. this parser is slightly different in its main loop compared to the old one, and may accidentally introduce bugs, though tests seem to be working fine
 2. there are potentially more places where description can be written into our fasta files (reconstruction and export)
 3. the easiest was to keep the `.desc` in output JSON (it is serialized as soon as the field added into the struct), but we could remove if it is not needed
 4. should desc participate in the generation of node ids?

For 3 and 4 I vote "keep" and "yes" because name+desc could provide a more unique identifier rather than name only. Without a desc it might not be possible to distinguish between the sequences in the outputs having the same name, if users rely on descriptions or have a malformed fasta with spaces in the name.