refactor: avoid panic
We want to avoid panics and stack unwinding as much as possible - the errors produced by panics lack context and usually are not very informative. They are also ugly and basically are crashes.

In order to handle the error properly, here I remove the call to `.expect()`, return a `Result` and propagate it up the call stack. Certain functions upper in the stack become fallible in the process, so I added error handling there as well.

I believe that `.unwrap()` in tests are fine, because that's not use-facing and they panic on assertions anyways.

The `is_anchor.then(...)` snippet has become a bit complicated and hard to read (probably suggested by Clippy). Can use some more love and refactoring. Perhaps a new named function can be extracted to replace the anonymous closure.