feat: add `--extra-band-width` and `--max-alignment-attempts`
This exposes previously hardcoded Nextclade-related params`EXTRA_BANDWIDTH` and `MAX_ALIGNMENT_ATTEMPTS` as CLI args of `pangraph build` command.

`max_alignment_attempts`is already a param of Nextclade, so it gets merged into the `NextalignParams`.

However, Nextclade is unaware of `extra_band_width` gets added to the band width directly. We may or may not backport this to Nextclade.